Move edges to child nodes
I noticed that currently, ELK will not make room for edges in compound nodes, because the edges exist on the parent. This change updates that, so that ELK will make room for the edges in a compound node.

However, I could not figure out a way to get it to make room for edges from inside a compound node to outside of it.