Add support for per node layout options
This adds support for setting layout options per node. They are copied from the `layoutOptions` key in the data.

This is useful for setting options like the partition (https://www.eclipse.org/elk/reference/options/org-eclipse-elk-partitioning-partition.html) which are meant to be set per node, to customize how each node is laid out.

If the layoutOptions is not set in the data, it will just be set to undefined to retain the status quo behavior.